deploy/requirements.txt
Flask
paho-mqtt
waitress